+++ src/mpcorbfile/__init__.py
@@ -2,5 +2,5 @@
 
 __author__ = """Nachoplus"""
 __email__ = "mas.ignacio@gmail.com"
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 from .mpcorbfile import mpcorb_file