+++ HISTORY.rst
@@ -2,6 +2,16 @@
 History
 =======
 
+0.3.2 (2025-03-04)
+------------------
+
+* Add calculated fields (expanded designation,orbit type, date of discovery from name)
+* fn to add bodies to REBOUND simulation
+* Added examples notebook for use with pyephem and REBOUND
+* get_chunk fn to divide bodies in packet for multiprocessing
+* Bug Fix
+
+
 0.1.0 (2025-03-04)
 ------------------